backend/pyproject.toml
[project]
name = "scp-world-backend"
version = "0.1.0"
description = "SCP World — FastAPI Backend with RAG, SSE Streaming, and Google OAuth 2.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "httpx>=0.28.0",
    "google-auth>=2.38.0",
    "requests>=2.32.0",
    "python-multipart>=0.0.20",
    "sentence-transformers>=3.0",
    "transformers>=4.0",
    "torch>=2.0",
    "google-cloud-firestore>=2.26.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "httpx>=0.28.0",
    "ruff>=0.9.0",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]
